- Moved the gigantic 10 000 line DrawFallBack text function from the SynthGlobals to its own file, in order to make the former easier to read/parse information from.
